\name{ccole-package}
\alias{ccole-package}
\alias{ccole}
\docType{package}
\title{
\packageTitle{ccole}
}
\description{
\packageDescription{ccole}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ccole}
\packageIndices{ccole}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ccole}

Maintainer: \packageMaintainer{ccole}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
