% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_con.R
\name{sql_con}
\alias{sql_con}
\title{Quick connect to SQLite db}
\usage{
sql_con(db_name = "db", drv = "SQLite()", file = NULL)
}
\arguments{
\item{db_name}{The db will be assigned to this. Defaults to db.}

\item{drv}{Type of connection}

\item{file}{Relative or absolute path to db.}
}
\description{
Quick connect to SQLite db
}

