% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2df.R
\name{tbl2df}
\alias{tbl2df}
\title{SQLite table to data frame}
\usage{
tbl2df(table = NULL, con = db)
}
\arguments{
\item{table}{Name of table in con}

\item{con}{Connection. Defaults to db.}
}
\description{
SQLite table to data frame
}

